Move NamedBeanType to package jmri.jmrit.logixng
