+++ java/src/jmri/jmrit/logixng/actions/swing/ActionListenOnBeansSwing.java
@@ -12,7 +12,7 @@ import javax.swing.table.TableColumn;
 import jmri.InstanceManager;
 import jmri.jmrit.logixng.*;
 import jmri.jmrit.logixng.actions.ActionListenOnBeans;
-import jmri.jmrit.logixng.actions.NamedBeanType;
+import jmri.jmrit.logixng.NamedBeanType;
 import jmri.jmrit.logixng.actions.ActionListenOnBeans.NamedBeanReference;
 import jmri.util.table.ButtonEditor;
 import jmri.util.table.ButtonRenderer;
