+++ java/src/jmri/jmrit/logixng/actions/swing/ListenOnBeansTableModel.java
@@ -12,7 +12,7 @@ import javax.swing.table.TableCellEditor;
 
 import jmri.Manager;
 import jmri.NamedBean;
-import jmri.jmrit.logixng.actions.NamedBeanType;
+import jmri.jmrit.logixng.NamedBeanType;
 import jmri.jmrit.logixng.actions.ActionListenOnBeans.NamedBeanReference;
 import jmri.swing.NamedBeanComboBox;
 import jmri.util.swing.JComboBoxUtil;
